file-format-detection/requirements.txt
cffi==1.14.5
xattr==0.9.7
python-magic==0.4.24
